:truck: Move old dashboard components
