+++ src/app/onion/dashboard/components/dashboard-item/dashboard-item.component.html
