+++ src/app/onion/dashboard/components/dashboard-item/dashboard-item.component.spec.ts
@@ -9,7 +9,7 @@ import { HttpClientModule } from '@angular/common/http';
 import { CookieModule } from 'ngx-cookie';
 import { CollectionService } from 'app/core/collection.service';
 import { LearningObject, User } from '@entity';
-import { DashboardLearningObject } from '../../old-dashboard.component';
+import { DashboardLearningObject } from '../../../old-dashboard/old-dashboard.component';
 import { TipDirective } from 'app/shared/tooltips/tip.directive';
 
 describe('DashboardItemComponent', () => {
