+++ src/app/onion/dashboard/components/side-panel/panel.directive.ts
