+++ src/app/onion/dashboard/components/side-panel/side-panel-viewer/side-panel-viewer.component.scss
