+++ src/app/onion/dashboard/components/side-panel/side-panel.module.ts
